cleanup

+++ meta_match.c
@@ -103,7 +103,7 @@ meta_regex_match(MetaRegex *regex, uint8 *input, int32 input_len, int64 nmatch,
         }
 
         for (int32 j = 0;; j += 1) {
-            uint8 b = (uint8)input[j];
+            uint8 b = input[j];
             int32 bit_match = (regex->fastmap[b >> 3] & (1 << (b & 7)));
 
             if (bit_match || regex->can_be_null) {
@@ -129,7 +129,7 @@ meta_regex_match(MetaRegex *regex, uint8 *input, int32 input_len, int64 nmatch,
         }
 
         for (int32 j = 0;; j += 1) {
-            uint8 b = (uint8)input[j];
+            uint8 b = input[j];
             int32 bit_match = (regex->fastmap[b >> 3] & (1 << (b & 7)));
 
             if (bit_match || regex->can_be_null) {
@@ -155,7 +155,7 @@ meta_regex_match(MetaRegex *regex, uint8 *input, int32 input_len, int64 nmatch,
         }
 
         for (int32 j = 0;; j += 1) {
-            uint8 b = (uint8)input[j];
+            uint8 b = input[j];
             int32 bit_match = (regex->fastmap[b >> 3] & (1 << (b & 7)));
 
             if (bit_match || regex->can_be_null) {
